refactor: sessionStatus 리팩터링
