--- src/main/java/wlsh/project/intervai/interview/domain/SessionStatus.java
@@ -1,6 +1,0 @@
-package wlsh.project.intervai.interview.domain;
-
-public enum SessionStatus {
-    IN_PROGRESS,
-    COMPLETED
-}
