+++ src/main/java/wlsh/project/intervai/interview/presentation/dto/InterviewSummaryResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import wlsh.project.intervai.interview.domain.Difficulty;
 import wlsh.project.intervai.interview.domain.InterviewSummary;
 import wlsh.project.intervai.interview.domain.InterviewType;
-import wlsh.project.intervai.interview.domain.SessionStatus;
+import wlsh.project.intervai.session.domain.SessionStatus;
 
 public record InterviewSummaryResponse(
         Long id,
