+++ src/main/java/wlsh/project/intervai/session/application/InterviewSessionValidator.java
@@ -2,28 +2,33 @@ package wlsh.project.intervai.session.application;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import wlsh.project.intervai.common.entity.EntityStatus;
 import wlsh.project.intervai.common.exception.CustomException;
 import wlsh.project.intervai.common.exception.ErrorCode;
-import wlsh.project.intervai.interview.application.InterviewFinder;
-import wlsh.project.intervai.interview.domain.Interview;
+import wlsh.project.intervai.interview.infra.InterviewEntity;
+import wlsh.project.intervai.interview.infra.InterviewRepository;
 import wlsh.project.intervai.session.infra.InterviewSessionEntity;
+import wlsh.project.intervai.session.infra.InterviewSessionRepository;
 
 @Component
 @RequiredArgsConstructor
 public class InterviewSessionValidator {
 
-    private final InterviewFinder interviewFinder;
-    private final InterviewSessionFinder interviewSessionFinder;
+    private final InterviewRepository interviewRepository;
+    private final InterviewSessionRepository interviewSessionRepository;
 
     public void validateInterviewOwner(Long interviewId, Long userId) {
-        Interview interview = interviewFinder.find(interviewId);
-        if (!interview.getUserId().equals(userId)) {
+        InterviewEntity entity = interviewRepository.findByIdAndStatus(interviewId, EntityStatus.ACTIVE)
+                .orElseThrow(() -> new CustomException(ErrorCode.INTERVIEW_NOT_FOUND));
+        if (!entity.isOwner(userId)) {
             throw new CustomException(ErrorCode.INTERVIEW_ACCESS_DENIED);
         }
     }
 
     public void validateSessionInProgress(Long interviewId) {
-        InterviewSessionEntity entity = interviewSessionFinder.getEntityByInterviewId(interviewId);
+        InterviewSessionEntity entity = interviewSessionRepository
+                .findByInterviewIdAndStatus(interviewId, EntityStatus.ACTIVE)
+                .orElseThrow(() -> new CustomException(ErrorCode.SESSION_NOT_FOUND));
         if (!entity.isInProgress()) {
             throw new CustomException(ErrorCode.SESSION_ALREADY_COMPLETED);
         }
