+++ src/main/java/wlsh/project/intervai/session/domain/InterviewSession.java
@@ -9,12 +9,12 @@ public class InterviewSession {
     private final Long id;
     private final Long interviewId;
     private final Long userId;
-    private final InterviewSessionStatus sessionStatus;
+    private final SessionStatus sessionStatus;
     private final int currentMainQuestionIdx;
     private final int followUpCount;
     private final LocalDateTime completedAt;
 
-    private InterviewSession(Long id, Long interviewId, Long userId, InterviewSessionStatus sessionStatus,
+    private InterviewSession(Long id, Long interviewId, Long userId, SessionStatus sessionStatus,
                              int currentMainQuestionIdx, int followUpCount, LocalDateTime completedAt) {
         this.id = id;
         this.interviewId = interviewId;
@@ -26,10 +26,10 @@ public class InterviewSession {
     }
 
     public static InterviewSession create(Long interviewId, Long userId) {
-        return new InterviewSession(null, interviewId, userId, InterviewSessionStatus.IN_PROGRESS, 0, 0, null);
+        return new InterviewSession(null, interviewId, userId, SessionStatus.IN_PROGRESS, 0, 0, null);
     }
 
-    public static InterviewSession of(Long id, Long interviewId, Long userId, InterviewSessionStatus sessionStatus,
+    public static InterviewSession of(Long id, Long interviewId, Long userId, SessionStatus sessionStatus,
                                       int currentMainQuestionIdx, int followUpCount, LocalDateTime completedAt) {
         return new InterviewSession(id, interviewId, userId, sessionStatus, currentMainQuestionIdx, followUpCount, completedAt);
     }
