+++ src/test/java/wlsh/project/intervai/interview/presentation/InterviewControllerTest.java
@@ -23,10 +23,10 @@ import wlsh.project.intervai.interview.domain.Interview;
 import wlsh.project.intervai.interview.domain.InterviewSummary;
 import wlsh.project.intervai.interview.domain.InterviewType;
 import wlsh.project.intervai.interview.domain.InterviewerTone;
-import wlsh.project.intervai.interview.domain.SessionStatus;
 import wlsh.project.intervai.interview.presentation.dto.CreateInterviewRequest;
 import wlsh.project.intervai.interview.presentation.dto.CsSubjectRequest;
 import wlsh.project.intervai.session.application.InterviewSessionService;
+import wlsh.project.intervai.session.domain.SessionStatus;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
